refactor(dropzone-image): move static and editor files to root of plugin
