+++ packages/editor/src/editor-integration/create-renderers.tsx
@@ -7,7 +7,7 @@ import { AnchorStaticRenderer } from '@editor/plugins/anchor/static'
 import { ArticleStaticRenderer } from '@editor/plugins/article/static'
 import { BlanksExerciseStaticRenderer } from '@editor/plugins/blanks-exercise/static'
 import { BoxStaticRenderer } from '@editor/plugins/box/static'
-import { DropzoneImageStaticRenderer } from '@editor/plugins/dropzone-image/components/static/static'
+import { DropzoneImageStaticRenderer } from '@editor/plugins/dropzone-image/static'
 import { EquationsStaticRenderer } from '@editor/plugins/equations/static'
 import { ExerciseStaticRenderer } from '@editor/plugins/exercise/static'
 import { GeogebraStaticRenderer } from '@editor/plugins/geogebra/static'
